+++ CHANGELOG.md
@@ -274,3 +274,4 @@
 - 2024-01-31: Refactor bettercontact, README, daemon, lib, +5
 - 2024-02-01: Impl bettercontact, discovery, LICENSE, lib, +3 query
 - 2024-02-01: Wire bettercontact, users, campaign, lib, +3
+- 2024-02-03: Async mod, follow up, qualify, find email, +5